adfa
afasdf

+++ 새 폴더/DX_1600/DX_1600/Object/CupHead/Gun/ChargeGun.h
@@ -18,5 +18,5 @@ private:
 
 	vector<shared_ptr<class ChargingBullet>> _Cbullets;
 
-	int _chargingCount = 0;
+	float _chargingCount = 0.0f;
 };
